requirements.txt
Flask==3.1.1
Flask-Mail==0.10.0
gunicorn==23.0.0
aiohttp==3.12.13
python-binance==1.0.29
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.41
dateparser==1.2.2
regex==2024.11.6
pytz==2025.2
tzlocal==5.3.1
